new Major version 1.delete funds from account 2.add fuelRate into block header
Former-commit-id: 25927b051ac27ebe42d02f00ce29c9ca95993301

+++ src/miner.h
@@ -71,7 +71,7 @@ bool CheckWork(CBlock* pblock, CWallet& wallet);
 /** Base sha256 mining transform */
 void SHA256Transform(void* pstate, void* pinput, const void* pinit);
 /** Get burn element */
-uint64_t GetElementForBurn(CBlockIndex *pindex);
+int GetElementForBurn(CBlockIndex *pindex);
 
 void GetPriorityTx(vector<TxPriority> &vecPriority);
 //extern double dHashesPerSec;
